Modularize speaking screen functions Changed files to make speaking screen composables reusable for the offline mode screens.
Files changed:
OfflineRecordingScreen - To include the reused composables
OfflineScreen - To change the downloaded image to an icon
Speaking - To include function calls instead of the raw composables
SpeakingFunctions - A new file which contains functions which can be called to reuse the composables
